Hardened Rust node: u64 timestamps, peer persistence, CHECO consensus, fraud propagation, STUN, deployment
Major changes across all four crates (trustchain-core, trustchain-transport, trustchain-node, trustchain-wasm):

Timestamp migration (f64 → u64 milliseconds):
- HalfBlock.timestamp is now u64 (ms since epoch), eliminating f64 rounding issues
- now_timestamp() → now_timestamp_ms(), all tests updated
- PeerRecord uses last_seen_unix_ms for persistence compatibility

Peer persistence:
- SQLite peers table with save/load/remove, WAL mode for concurrent readers
- BlockStore trait extended with save_peer/load_peers/remove_stale_peer
- Node loads persisted peers on startup, saves each gossip tick

Protocol audit fixes:
- Removed dead-code && guards in cross-link pubkey check and double-countersign detection (halfblock.rs)
- CHECO availability: stale checkpoint only rejects our_seq > claimed_seq, not the reverse

QUIC hardening:
- Per-IP rate limiter (configurable max_connections_per_ip_per_sec, default 20)
- Connection reuse cache (active_connections HashMap, transparent reconnect on dead connections)

Fraud propagation + trust penalty:
- FraudProof QUIC message type with TTL relay and dedup
- compute_trust() returns hard 0.0 for agents with proven double-spend
- CHECO message types: CheckpointProposal, CheckpointVote, CheckpointFinalized

CHECO consensus network rounds:
- checkpoint_loop in node runtime, facilitator election, dynamic known_peers
- CheckpointWire serialization for over-the-wire checkpoint exchange

STUN NAT traversal:
- Minimal STUN client (stun.rs), XOR-MAPPED-ADDRESS parsing
- advertise_addr config + --advertise sidecar flag
- Self-registration uses public addr (STUN-derived or explicit), not 127.0.0.1

Proxy improvements:
- Per-peer try_lock coalescing (sequential interactions per peer pair per paper)
- Atomic handshake (protocol lock held across QUIC round-trip)
- POST /peers HTTP endpoint for runtime peer registration

Node runtime:
- Graceful shutdown via tokio::signal::ctrl_c
- Config: max_connections_per_ip_per_sec, checkpoint_interval_secs, stun_server, advertise_addr
- effective_bootstrap_nodes() with built-in seed node fallback

Deployment artifacts:
- Dockerfile (multi-stage build with deploy/docker-node.toml baked in)
- systemd service unit (deploy/trustchain.service)
- GitHub Actions CI/CD workflow (.github/workflows/release.yml)
- install.sh one-liner

Test suite: 166 tests passing (93 core + 8 stress + 7 stress2 + 5 stress3 + 4 stress4 + 7 system_tests + 42 transport)

+++ trustchain-rs/trustchain-core/src/protocol.rs
@@ -55,7 +55,7 @@ impl<S: BlockStore> TrustChainProtocol<S> {
         &mut self,
         counterparty_pubkey: &str,
         transaction: serde_json::Value,
-        timestamp: Option<f64>,
+        timestamp: Option<u64>,
     ) -> Result<HalfBlock> {
         let seq = self.store.get_latest_seq(&self.pubkey())? + 1;
         let prev_hash = self.store.get_head_hash(&self.pubkey())?;
@@ -135,7 +135,7 @@ impl<S: BlockStore> TrustChainProtocol<S> {
     pub fn create_agreement(
         &mut self,
         proposal: &HalfBlock,
-        timestamp: Option<f64>,
+        timestamp: Option<u64>,
     ) -> Result<HalfBlock> {
         // Must be a proposal block.
         if !proposal.is_proposal() {
@@ -360,7 +360,7 @@ mod tests {
         let bob_key = Identity::from_bytes(&[2u8; 32]).pubkey_hex();
 
         let proposal = alice
-            .create_proposal(&bob_key, serde_json::json!({"service": "compute"}), Some(1000.0))
+            .create_proposal(&bob_key, serde_json::json!({"service": "compute"}), Some(1000))
             .unwrap();
 
         assert_eq!(proposal.sequence_number, 1);
@@ -379,7 +379,7 @@ mod tests {
             .create_proposal(
                 &bob.pubkey(),
                 serde_json::json!({"service": "compute", "amount": 100}),
-                Some(1000.0),
+                Some(1000),
             )
             .unwrap();
 
@@ -387,7 +387,7 @@ mod tests {
         assert!(bob.receive_proposal(&proposal).unwrap());
 
         // Bob creates agreement.
-        let agreement = bob.create_agreement(&proposal, Some(1001.0)).unwrap();
+        let agreement = bob.create_agreement(&proposal, Some(1001)).unwrap();
         assert!(agreement.is_agreement());
         assert_eq!(agreement.link_public_key, alice.pubkey());
         assert_eq!(agreement.link_sequence_number, proposal.sequence_number);
@@ -410,12 +410,12 @@ mod tests {
                 .create_proposal(
                     &bob.pubkey(),
                     serde_json::json!({"round": i}),
-                    Some(1000.0 + i as f64 * 2.0),
+                    Some(1000 + i as u64 * 2),
                 )
                 .unwrap();
 
             bob.receive_proposal(&proposal).unwrap();
-            let agreement = bob.create_agreement(&proposal, Some(1001.0 + i as f64 * 2.0)).unwrap();
+            let agreement = bob.create_agreement(&proposal, Some(1001 + i as u64 * 2)).unwrap();
             alice.receive_agreement(&agreement).unwrap();
         }
 
@@ -436,7 +436,7 @@ mod tests {
 
         // Alice proposes to Charlie, not Bob.
         let proposal = alice
-            .create_proposal(&charlie_key, serde_json::json!({}), Some(1000.0))
+            .create_proposal(&charlie_key, serde_json::json!({}), Some(1000))
             .unwrap();
 
         // Bob should reject it.
@@ -453,11 +453,11 @@ mod tests {
         );
 
         let proposal = alice
-            .create_proposal(&bob.pubkey(), serde_json::json!({}), Some(1000.0))
+            .create_proposal(&bob.pubkey(), serde_json::json!({}), Some(1000))
             .unwrap();
 
         bob.receive_proposal(&proposal).unwrap();
-        let agreement = bob.create_agreement(&proposal, Some(1001.0)).unwrap();
+        let agreement = bob.create_agreement(&proposal, Some(1001)).unwrap();
 
         // Charlie should reject the agreement (it links to Alice, not Charlie).
         let result = charlie.receive_agreement(&agreement);
@@ -469,12 +469,12 @@ mod tests {
         let (mut alice, mut bob) = make_protocol();
 
         let proposal = alice
-            .create_proposal(&bob.pubkey(), serde_json::json!({}), Some(1000.0))
+            .create_proposal(&bob.pubkey(), serde_json::json!({}), Some(1000))
             .unwrap();
 
         // Bob creates agreement without first receiving the proposal.
         bob.receive_proposal(&proposal).unwrap();
-        let agreement = bob.create_agreement(&proposal, Some(1001.0)).unwrap();
+        let agreement = bob.create_agreement(&proposal, Some(1001)).unwrap();
 
         // Create a fresh Alice that doesn't have the proposal in store.
         let mut alice2 = TrustChainProtocol::new(
@@ -493,7 +493,7 @@ mod tests {
             .create_proposal(
                 &bob.pubkey(),
                 serde_json::json!({"service": "compute"}),
-                Some(1000.0),
+                Some(1000),
             )
             .unwrap();
 
@@ -504,7 +504,7 @@ mod tests {
         tampered_proposal.transaction = serde_json::json!({"service": "FAKE"});
         // Bob would create agreement with wrong transaction — but create_agreement
         // copies from proposal, so we'd need to tamper post-creation.
-        let agreement = bob.create_agreement(&proposal, Some(1001.0)).unwrap();
+        let agreement = bob.create_agreement(&proposal, Some(1001)).unwrap();
 
         // Tamper the agreement after creation.
         let mut tampered_agreement = agreement.clone();
@@ -521,10 +521,10 @@ mod tests {
 
         for i in 0..3 {
             let proposal = alice
-                .create_proposal(&bob.pubkey(), serde_json::json!({"i": i}), Some(1000.0 + i as f64))
+                .create_proposal(&bob.pubkey(), serde_json::json!({"i": i}), Some(1000 + i as u64))
                 .unwrap();
             bob.receive_proposal(&proposal).unwrap();
-            let agreement = bob.create_agreement(&proposal, Some(1001.0 + i as f64)).unwrap();
+            let agreement = bob.create_agreement(&proposal, Some(1001 + i as u64)).unwrap();
             alice.receive_agreement(&agreement).unwrap();
         }
 
@@ -542,10 +542,10 @@ mod tests {
         let (mut alice, mut bob) = make_protocol();
 
         let proposal = alice
-            .create_proposal(&bob.pubkey(), serde_json::json!({}), Some(1000.0))
+            .create_proposal(&bob.pubkey(), serde_json::json!({}), Some(1000))
             .unwrap();
         bob.receive_proposal(&proposal).unwrap();
-        let agreement = bob.create_agreement(&proposal, Some(1001.0)).unwrap();
+        let agreement = bob.create_agreement(&proposal, Some(1001)).unwrap();
         alice.receive_agreement(&agreement).unwrap();
 
         assert!(alice.validate_chain(&alice.pubkey()).unwrap());
@@ -558,12 +558,12 @@ mod tests {
 
         for i in 1..=3 {
             let proposal = alice
-                .create_proposal(&bob.pubkey(), serde_json::json!({"i": i}), Some(1000.0 + i as f64))
+                .create_proposal(&bob.pubkey(), serde_json::json!({"i": i}), Some(1000 + i as u64))
                 .unwrap();
             assert_eq!(proposal.sequence_number, i);
 
             bob.receive_proposal(&proposal).unwrap();
-            let agreement = bob.create_agreement(&proposal, Some(1001.0 + i as f64)).unwrap();
+            let agreement = bob.create_agreement(&proposal, Some(1001 + i as u64)).unwrap();
             assert_eq!(agreement.sequence_number, i);
 
             alice.receive_agreement(&agreement).unwrap();
@@ -575,16 +575,16 @@ mod tests {
         let (mut alice, mut bob) = make_protocol();
 
         let p1 = alice
-            .create_proposal(&bob.pubkey(), serde_json::json!({"i": 1}), Some(1000.0))
+            .create_proposal(&bob.pubkey(), serde_json::json!({"i": 1}), Some(1000))
             .unwrap();
         assert_eq!(p1.previous_hash, GENESIS_HASH);
 
         bob.receive_proposal(&p1).unwrap();
-        let a1 = bob.create_agreement(&p1, Some(1001.0)).unwrap();
+        let a1 = bob.create_agreement(&p1, Some(1001)).unwrap();
         alice.receive_agreement(&a1).unwrap();
 
         let p2 = alice
-            .create_proposal(&bob.pubkey(), serde_json::json!({"i": 2}), Some(1002.0))
+            .create_proposal(&bob.pubkey(), serde_json::json!({"i": 2}), Some(1002))
             .unwrap();
         assert_eq!(p2.previous_hash, p1.block_hash);
     }
